refactor: shared utilities
